realestateapi/requirements.txt
# Django Web Framework
Django==5.0.6
asgiref==3.7.0
django-environ==0.9.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
# Django Cors Headers - Used to enable CORS headers in API responses, and allow requests to be made to your API server from other origins.
django-cors-headers==4.3.1

# Django Rest Framework - Used to design API logic
djangorestframework==3.15.0

# MySQL Client - Used as an interface to connect Django application to the MySQL server.
#mysqlclient==2.0.1